Refactor PublicKey to another file; Rename to ASN1PublicKey
